Update submit event info (#453)
* clean up tests

* remove attributes from model

* remove metadata_version from submit event info
